arc score rework

+++ src/utils/dataStructure/arc.h
@@ -20,7 +20,7 @@
 typedef struct t_arc_p{
     int candidat_gagnant;
     int candidat_perdant;
-    float score;
+    int score;
 }arc;
 
 typedef struct s_arc{
@@ -35,7 +35,7 @@ typedef struct s_arc{
  * @param score
  * @return arc* l'arc créé
  */
-arc* arc_create(int candidat_gagnant, int candidat_perdant, float score);
+arc* arc_create(int candidat_gagnant, int candidat_perdant, int score);
 
 /**
  * @brief Fonction qui permet d'initialiser et de completer la liste des arcs à partir de la matrice de duel.
